archive_docker_server/archive_docker_client/requirements.txt
grpcio
grpcio-tools
kafka-python
clickhouse-connect